Shadow plug-in now uses jar file to compile.
Previously, the Shadow plug-in used shadowc, which had to be available
on a path.  Now, the plug-in will use whatever jar is specified in its
preferences pages, allowing for the case when shadowc is not installed
or when the plug-in should be used with a specific shadow.jar file
(which does not need to be named shadow.jar).

Also added error message when no jar file is specified.