Changed Shadow plug-in to include shadow.jar.
Doing so allows for much cleaner code, directly using classes and
methods from the compiler instead of using reflection.

This change is also updates to antlr, since using the new shadow.jar
with the old plug-in would give strange output for the outline.

A few other small changes were made, including updates to the list of
reserved words (send and spawn), addition of icons for protected and
private constants in the outline, and a Source menu with a Correct
Indentation command.

Since shadow.jar is now internal to the plug-in, the preferences for the
plug-in no longer allow the user to specify the shadow.jar being used.
However, the preferences require the user to specify the location of a
shadow.xml configuration file which gives the location of system
imports.  Otherwise, the plug-in doesn't have a standard library to use
when type-checking.

The plug-in also looks for a SHADOW_HOME environment variable, which it
will use for a default location of the compiler executable. 

At present, compiling a Shadow file through the plug-in crashes Eclipse
and the JVM. I don't know why.  The compilation succeeds, and it's
possible to run the file through the