Updated launch of programs from plug-in.
Now, command line arguments to the compiler will be parsed correctly,
even if there are arguments delimited by quotation marks.

The compiler figures out if a custom output file is specified and runs
that.

Fixes have also been made to how the custom compiler is specified in the
run configuration.  The field can now be left empty, defaulting to
shadowc.