Allowed multiple parse errors to show.
Doing so required reworking some of the ways that parse errors are
reported by the compiler.

Now, the outline will show whatever the parser has been able to come up
with instead of the line with the error on it.

There's unfortunately a little redundancy, since the type-checker will
do a parse first, using the same machinery as the outline.  It would be
nice if one parse could be done instead of two, but Antlr is pretty
fast.

On an unrelated note, I can't get the grammar to recognize something
like:

int value = 5;
int other = --value;

In Shadow, that should be legal (though useless), setting other to the
negation of the negation of value.  However, the parser doesn't like it,
permitting only - -value.  Equally strange, the boolean equivalent of
!!choice works fine.