Added error checking to plug-in preference page.
Also, changed all the plug-in naming from org.shadow-language to
shadow.plugin.  In some ways, that's a move away from convention, but it
makes the plug-in more consistent.  Also, it's impossible to have
everything match since org.shadow-language is not a legal Java package
name.

I wish we had a better domain.  Unfortunately, shadow.org and
shadow.codes are both claimed (though apparently not used for a web
page).